Python POC: add tests about error handling
