+++ openeo_compliance_tests/openeo_compliance_tests/test_general.py
@@ -1,8 +1,6 @@
-import _pytest.python
 import pytest
-import requests
 
-from openeo_compliance_tests.helpers import ApiSchemaValidator, ApiClient, ResponseNotInSchema, get_api_version
+from openeo_compliance_tests.helpers import ApiSchemaValidator, ApiClient, ResponseNotInSchema
 
 
 @pytest.mark.parametrize("path", [
@@ -51,32 +49,3 @@ def test_well_known_openeo(client, schema, api_version):
     client = ApiClient(backend=client.domain)
     test_get_generic(client=client, schema=schema, api_version=api_version, path=path)
 
-
-def test_get_unauthorized(client: ApiClient, unauthorized_get_path: str):
-    """
-    Check that paths that require authentication indeed return with a 4XX HTTP code
-    """
-    response = client.get(unauthorized_get_path)
-    if response.status_code == requests.codes.not_found:
-        pytest.skip('{p} not found'.format(p=unauthorized_get_path))
-    assert response.status_code in (
-        requests.codes.unauthorized,  # HTTP status code 401
-        requests.codes.forbidden,  # HTTP status code 403
-        requests.codes.not_allowed,  # HTTP status code 405
-    )
-
-
-def pytest_generate_tests(metafunc: _pytest.python.Metafunc):
-    """
-    Pytest hook for custom test parametrization
-    """
-    if 'unauthorized_get_path' in metafunc.fixturenames:
-        schema = ApiSchemaValidator.from_version(get_api_version(metafunc.config))
-        # Search for get requests which require authentication
-        paths = []
-        for path in schema.get_paths():
-            if '{' not in path:
-                path_get_schema = schema.get_path_schema(path).get('get', {})
-                if 'security' in path_get_schema and all(s != {} for s in path_get_schema['security']):
-                    paths.append(path)
-        metafunc.parametrize('unauthorized_get_path', paths)
